Add eval_id to few structures
We'll need them in the next commit.

This also unifies type of eval_id to u64.

+++ src/cli.rs
@@ -20,7 +20,7 @@ pub enum Commands {
 #[derive(Args)]
 pub struct EvalArgs {
     /// Evaluation ID to analyze
-    pub eval_id: usize,
+    pub eval_id: u64,
 
     /// Print statistics about evaluation failures
     #[arg(long)]
